Added Sidebar skeleton on getting jwt token, SideBar code refactor
