+++ src/components/layout/index.ts
@@ -1,3 +1,4 @@
 export { Footer } from './Footer'
 export { Header } from './Header'
 export { AppLayout } from './AppLayout'
+export { SideNav } from './SideNav'
