+++ src/config/constants.ts
@@ -1,5 +1,6 @@
 export const DEFAULT_LANG = 'it'
 export const STORAGE_KEY_SESSION_TOKEN = 'token'
+export const SIDENAV_WIDTH = 340
 
 export const MOCK_TOKEN = import.meta.env.REACT_APP_MOCK_TOKEN
